Added padding to favorite recipes and my meal plans screen
